1.Spell check and minor code adjustment before migrate to Linux
2.Bug fixed: miss to release mutex lock in asynchronous mode of send

3.Bug fixed: should NotifyToAccept if persist c/w EoT received

4.Minor refactor in the Disconnect function

5.Protocol glitch fixed: add SHUT_REQUESTED state

6.Protocol glitch fixed: should not use a field
  in encrypted extension header as the salt that decryption depends on
  There's a tradeoff between encrypting the extension header or not.

7.Protocol optimized: giving the sequence number of the packet
  that was received latest explicitly for calibrating acknowledgement delay

8.Fixed: should calculate the first RTT as soon as possible
  And eliminatate delay caused by calling back ULA
- OnConnectRequestAck (UpdateRTT)
- OnGetPersist
- OnGetNulCommit
- OnGetMultiply & RespondToMultiply (record time of receive first packet)

9.Fixed: timeDelta is completely transparent to the initiator side,
  And need not be encoded in network byte order

10.Fixed: error on calculating advertised window size
  when out-of-band packet is sent

11.Optimize (and fix the bug of) garbage collection on allocating new socket in DLL

12.Test VPN based on FSP/TCP SOCKSv4 Server

+++ src/FSP_DLL/Multiplex.cpp
@@ -45,7 +45,7 @@
 //	which indicate some error has happened.
 //	See also SendStream, FinalizeSend
 DllExport
-FSPHANDLE FSPAPI MultiplyAndWrite(FSPHANDLE hFSP, PFSP_Context psp1, int8_t flag, NotifyOrReturn fp1)
+FSPHANDLE FSPAPI MultiplyAndWrite(FSPHANDLE hFSP, PFSP_Context psp1, unsigned flag, NotifyOrReturn fp1)
 {
 	CommandCloneConnect objCommand;
 	CSocketItemDl *p = CSocketItemDl::ToPrepareMultiply(hFSP, psp1, objCommand);
@@ -168,7 +168,7 @@ CSocketItemDl * LOCALAPI CSocketItemDl::ToPrepareMultiply(FSPHANDLE h, PFSP_Cont
 
 
 inline
-FSPHANDLE LOCALAPI CSocketItemDl::WriteOnMultiplied(CommandCloneConnect &objCommand, PFSP_Context psp1, int8_t flag, NotifyOrReturn fp1)
+FSPHANDLE LOCALAPI CSocketItemDl::WriteOnMultiplied(CommandCloneConnect &objCommand, PFSP_Context psp1, unsigned flag, NotifyOrReturn fp1)
 {
 	TestSetSendReturn(fp1);
 	if (psp1->welcome != NULL)
